refactor(runtime-host): trim dead export surface from barrels
Remove unconsumed server, client, protocol, and adapter barrel exports while moving package-internal consumers to their leaf modules.

Generated-by: Codex
