+++ packages/runtime-host/src/__tests__/authenticated-websocket.test.ts
@@ -18,10 +18,8 @@ import {
   RUNTIME_HOST_PROTOCOL_VERSION,
   type RequestFrame,
 } from '../protocol/index.js';
-import {
-  openRuntimeHostAccessAuthority,
-  startExecutionRuntimeHostService,
-} from '../server/index.js';
+import { openRuntimeHostAccessAuthority } from '../server/access-authority.js';
+import { startExecutionRuntimeHostService } from '../server/execution-service.js';
 import { authorizeRuntimeHostOperation } from '../server/connection-authority.js';
 
 const PROTOCOL = {
