+++ packages/runtime-host/src/adapter/index.ts
@@ -4,7 +4,6 @@ export {
   isRuntimeHostTerminalTurn,
   foldRuntimeHostAssistantDelta,
   projectRuntimeHostInteractionRequest,
-  type RuntimeHostProjectionUpdate,
   type RuntimeHostSessionProjectionSeed,
   type RuntimeHostTerminalTurn,
 } from './session-projector.js';
