+++ packages/runtime-host/src/server/session-catalog-coordinator.ts
@@ -1,3 +1,4 @@
+import { RuntimeHostProtocolError } from '../protocol/errors.js';
 import { createHash } from 'node:crypto';
 import { isModelExplicitlyUnsupportedForChat } from '@maka/core/model-catalog';
 import { thinkingVariantsForConnection } from '@maka/core/model-thinking';
@@ -37,7 +38,6 @@ import {
   SESSION_CATALOG_RESULT_MAX_BYTES,
   type OperationError,
   type OperationOutcome,
-  RuntimeHostProtocolError,
   type SessionCatalogFilter,
   type SessionCatalogItem,
   type SessionCatalogProjection,
